requirements.txt
selenium==4.15.0 
pytest==7.4.0 
allure-pytest==2.15.0 
pytest-rerunfailures==10.0 
pytest-cov==6.2.1 
